requirements.txt
amqp
asgiref
async-timeout
b5
beautifulsoup4
billiard
bs4
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
Django
django-cors-headers
django-extensions
django-filter
djangorestframework
drf-yasg
exceptiongroup
factory-boy
Faker
flower
humanize
idna
inflection
iniconfig
Jinja2
kombu
lxml
MarkupSafe
packaging
Pillow
pluggy
prometheus-client
prompt-toolkit
psycopg2-binary
PyJWT
pymemcache
pypng
pyTelegramBotAPI
pytest
pytest-django
python-dateutil
python-decouple
python-slugify
pytz
PyYAML
qrcode
redis
requests
six
soupsieve
sqlparse
termcolor
text-unidecode
tomli
tornado
typing_extensions
tzdata
uritemplate
urllib3
wcwidth
